cross-window fix, drag & drop added to test files
- In addEventHandler(), the check to see if the obj argument was valid
didn't work if the object was in an iframe. If obj is in a different
window, it isn't an instance of this window's Object(); it is an
instance of the iframe window's Object(). (I wasn't aware that they were
different until now.)

- Added to the test_events_max files to test drag & drop events, and
created a new file for use in the iframe.